requirements.txt
asgiref==3.8.1
dj-database-url==2.1.0
Django==5.1.1
djangorestframework==3.15.2
googleapis-common-protos==1.65.0
grpcio==1.66.2
grpcio-status==1.66.2
gunicorn==21.2.0
numpy==1.26.4
packaging==23.2
pandas==2.2.3
Pillow==10.1.0
protobuf==5.28.2
psycopg2-binary==2.9.9
py4j==0.10.9.7
pyarrow==17.0.0
pyspark==3.5.3
python-dateutil==2.9.0.post0
pytz==2024.2
six==1.16.0
sqlparse==0.4.4
typing_extensions==4.8.0
tzdata==2024.2
whitenoise==6.6.0
openai